+++ plugin_tooling/src/dtool/ast/definitions/AbstractFunctionDefinition.java
@@ -22,7 +22,6 @@ import melnorme.utilbox.collections.ArrayView;
 import melnorme.utilbox.core.CoreUtil;
 import melnorme.utilbox.misc.IteratorUtil;
 import dtool.ast.expressions.Expression;
-import dtool.ast.references.RefTemplateInstance;
 import dtool.ast.statements.IFunctionBody;
 import dtool.parser.common.Token;
 
@@ -86,11 +85,6 @@ public abstract class AbstractFunctionDefinition extends CommonDefinition
 	@Override
 	public NodeVector<ITemplateParameter> getTemplateParameters() {
 		return tplParams;
-	}
-	
-	@Override
-	public DefUnit cloneTemplateElement(RefTemplateInstance templateRef) {
-		return clone(this); /*FIXME: BUG here*/
 	}
 	
 	/* -----------------  ----------------- */