+++ documentation/ChangeLog.md
@@ -5,6 +5,7 @@
  * Fixed: Error parsing UDAS as function attributes (example: `int foo() @blah { }`) .
  * Fixed: removed parser errors about types being used in expressions. 
    * Fixes false errors occurring in valid code such as `auto x = int(223)`;
+ * Support for basic template instantiation in semantic engine.
 
 ### DDT 0.10.4
  * Fixed #88: Semantic operations (code complete, etc.) involving DUB packages with subpackages would always cause the `dub describe` to run and thus incur a significant slowdown.